Add metrics and restructure
