+++ asreviewcontrib/insights/__init__.py
@@ -1,20 +1,6 @@
-# Copyright 2020 The ASReview Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from asreviewcontrib.insights.recall import plot_recall
-from asreviewcontrib.insights.wss import plot_wss
-from asreviewcontrib.insights.erf import plot_erf
+from asreviewcontrib.insights.plot import plot_recall
+from asreviewcontrib.insights.plot import plot_wss
+from asreviewcontrib.insights.plot import plot_erf
 
 from ._version import get_versions
