+++ asreviewcontrib/insights/entrypoint.py
@@ -1,17 +1,3 @@
-# Copyright 2020 The ASReview Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
 import argparse
 import json
 import logging
@@ -22,10 +8,6 @@ import matplotlib.pyplot as plt
 from asreview.entry_points import BaseEntryPoint
 from asreview import open_state
 
-# from asreviewcontrib.visualization import Plot
-# from asreviewcontrib.visualization.quick import progression_plot
-# from asreviewcontrib.visualization.quick import limit_plot
-# from asreviewcontrib.visualization.quick import discovery_plot
 from asreviewcontrib.insights import plot_recall
 from asreviewcontrib.insights import plot_wss
 from asreviewcontrib.insights import plot_erf
