+++ docs/example_custom_title.py
@@ -1,7 +1,7 @@
 import matplotlib.pyplot as plt
 
 from asreview import open_state
-from asreviewcontrib.insights import plot_wss
+from asreviewcontrib.insights.plot import plot_wss
 
 with open_state("tests/asreview_files/sim_van_de_schoot_2017_1.asreview") as s:
